+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Current (in progress)
+
+- Nothing yet
+
 ## 2.2.2 (2025-03-14)
 
 - Add Dataservices views metrics as field factor and sort [#52](https://github.com/opendatateam/udata-search-service/pull/52)