+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current (in progress)
 
 - Add object id in search fields [#55](https://github.com/opendatateam/udata-search-service/pull/55)
+- Add resources ids and titles in search fields [#56](https://github.com/opendatateam/udata-search-service/pull/56)
 
 ## 2.2.2 (2025-03-14)