+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add malus score for archived reuses [#54](https://github.com/opendatateam/udata-search-service/pull/54)
 - Add object id in search fields [#55](https://github.com/opendatateam/udata-search-service/pull/55)
 - Add resources ids and titles in search fields [#56](https://github.com/opendatateam/udata-search-service/pull/56)
+- Upgrade deps [#57](https://github.com/opendatateam/udata-search-service/pull/57)
 
 ## 2.2.2 (2025-03-14)